+++ src/store/content.js
@@ -1,23 +1,65 @@
 export const actions = {
 	async getResources({ commit }, payload = {}) {
 		commit("setLoading", true);
-		const query = Object.assign({ $limit: 10 }, payload || {});
-		const res = await this.$axios.$get("/content/search", {
-			params: query,
-		});
-		commit("setResources", res);
-		commit("setLoading", false);
+		const query = {
+			$limit: 12,
+			$skip: 0,
+			...payload,
+		};
+		try {
+			const res = await this.$axios.$get("/edu-sharing", {
+				params: query,
+			});
+			commit("setResources", res);
+		} finally {
+			commit("setLoading", false);
+		}
 	},
 	async addResources({ commit }, payload = {}) {
 		commit("setLoading", true);
-		const query = payload || {};
-		const res = await this.$axios.$get("/content/search", {
-			params: query,
+		const res = await this.$axios.$get("/edu-sharing", {
+			params: payload,
 		});
 		commit("addResources", res);
 		commit("setLoading", false);
 	},
+	async getLessons({ commit }, payload) {
+		const params = {
+			courseId: payload,
+		};
+		const res = await this.$axios.$get("/lessons", { params });
+		commit("setLessons", res);
+	},
+
+	async addToLesson({ commit }, payload = { material: {} }) {
+		await this.$axios
+			.post(`/lessons/${payload.lessonId}/material`, payload.material)
+			.then((resp) => {
+				commit("addToLessonResult", resp);
+			})
+			.catch((error) => {
+				console.error(`addToLessonResult Error: ${error}, payload: ${payload}`);
+				commit("addToLessonResult", error.response);
+			});
+	},
+	async getResourceMetadata(context, id) {
+		return this.$axios.$get(`/edu-sharing/${id}`);
+	},
 };
+
+const initialState = () => ({
+	resources: {
+		total: 0,
+		limit: 0,
+		skip: 0,
+		data: [],
+	},
+	lessons: {
+		data: [],
+	},
+	loading: false,
+	addToLessonResult: {},
+});
 
 export const mutations = {
 	setResources(state, payload) {
@@ -27,22 +69,21 @@ export const mutations = {
 		payload.data.forEach((resource) => state.resources.data.push(resource));
 		state.resources = {
 			...state.resources,
-			limit: payload.limit,
-			skip: payload.skip,
-			total: payload.total,
+			pagination: payload.pagination,
 		};
 	},
-	setLoading(state, type) {
-		state.loading = type;
+	clearResources(state) {
+		state.resources = initialState().resources;
+	},
+	setLoading(state, status) {
+		state.loading = status;
+	},
+	setLessons(state, payload) {
+		state.lessons = payload;
+	},
+	addToLessonResult(state, payload) {
+		state.addToLessonResult = payload;
 	},
 };
 
-export const state = () => ({
-	resources: {
-		data: [],
-		limit: null,
-		skip: null,
-		total: null,
-	},
-	loading: false,
-});
+export const state = initialState();